+++ tests/test_data/output_grouped_single_line.py
@@ -39,5 +39,4 @@ from z import foo
 from ..othermodule import rainbows
 from .module import bar, foo
 
-
 # stuff here