+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+0.6.0 (2017-10-06)
+~~~~~~~~~~~~~~~~~~
+
+* Added support for sub-configurations when ``.importanize`` is found.
+* Added support for ``add_imports`` in configuration.
+
 0.5.3 (2017-06-06)
 ~~~~~~~~~~~~~~~~~~